requirements.txt
requests
bs4
pyrogram==1.4.16
tgcrypto
cloudscraper
python-dotenv
